+++ packages/adapter-typestyle/src/TypeStyleAesthetic.ts
@@ -1,4 +1,6 @@
-import Aesthetic, { AestheticOptions, ClassName, Ruleset, Sheet } from 'aesthetic';
+/* eslint-disable no-underscore-dangle */
+
+import Aesthetic, { purgeStyles, AestheticOptions, ClassName, Ruleset, Sheet } from 'aesthetic';
 import { TypeStyle } from 'typestyle';
 import { NativeBlock, ParsedBlock } from './types';
 
@@ -28,6 +30,19 @@ export default class TypeStyleAesthetic<Theme extends object> extends Aesthetic<
       .on('pseudo', this.handleNested)
       .on('selector', this.handleNested)
       .on('support', this.handleSupport);
+  }
+
+  flushStyles() {
+    this.typeStyle.forceRenderStyles();
+  }
+
+  purgeStyles() {
+    // @ts-ignore
+    const element: HTMLStyleElement | undefined = this.typeStyle._tag;
+
+    if (element) {
+      purgeStyles(element);
+    }
   }
 
   transformToClassName(styles: ParsedBlock[]): ClassName {