+++ packages/core/src/helpers/hasQueryCondition.ts
@@ -1,0 +1,5 @@
+const QUERY_GROUP = /\([-a-z]+:/giu;
+
+export default function hasQueryCondition(value: string): boolean {
+  return !!value.match(QUERY_GROUP);
+}