+++ packages/core/src/index.ts
@@ -8,9 +8,23 @@ import ClassNameAesthetic from './ClassNameAesthetic';
 import UnifiedSyntax from './UnifiedSyntax';
 import Ruleset from './Ruleset';
 import Sheet from './Sheet';
+import getFlushedStyles from './helpers/getFlushedStyles';
+import getStyleElements from './helpers/getStyleElements';
+import purgeStyles from './helpers/purgeStyles';
+import isRTL from './helpers/isRTL';
 
 export * from './types';
 
-export { AestheticOptions, ClassNameAesthetic, UnifiedSyntax, Ruleset, Sheet };
+export {
+  AestheticOptions,
+  ClassNameAesthetic,
+  UnifiedSyntax,
+  Ruleset,
+  Sheet,
+  getFlushedStyles,
+  getStyleElements,
+  isRTL,
+  purgeStyles,
+};
 
 export default Aesthetic;