+++ packages/react/src/useStylesFactory.ts
@@ -1,5 +1,5 @@
 import { useContext, useState, useLayoutEffect } from 'react';
-import Aesthetic, { ClassNameTransformer, StyleSheetDefinition, SheetMap } from 'aesthetic';
+import Aesthetic, { ClassNameTransformer, StyleSheetDefinition, SheetMap, isRTL } from 'aesthetic';
 import uuid from 'uuid/v4';
 import DirectionContext from './DirectionContext';
 
@@ -27,7 +27,7 @@ export default function useStylesFactory<
     });
 
     // Create a unique style sheet for this component
-    const options = { name: styleName, rtl: aesthetic.isRTL(dir) };
+    const options = { name: styleName, rtl: isRTL(dir) };
     const sheet = aesthetic.createStyleSheet(styleName, options);
 
     // Flush styles on mount